pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "graph-of-marks"
version = "1.0.1"
description = "Visual scene understanding pipeline with multi-model detection, segmentation, and scene graph generation"
readme = "README.md"
requires-python = ">=3.8"
license = "MIT"
license-files = ["LICENSE"]
authors = [
    {name = "DISI-UNIBO-NLP"}
]
keywords = [
    "computer-vision",
    "scene-graph",
    "object-detection",
    "segmentation",
    "visual-question-answering",
    "sam",
    "yolo",
    "clip",
    "deep-learning",
    "pytorch",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Scientific/Engineering :: Image Recognition",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Operating System :: POSIX :: Linux",
    "Operating System :: MacOS :: MacOS X",
    "Environment :: GPU :: NVIDIA CUDA",
]

dependencies = [
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "pillow>=9.0.0",
    "numpy>=1.24.0",
    "opencv-python>=4.8.0",
    "transformers>=4.36.0",
    "networkx>=3.1",
    "matplotlib>=3.5.0",
    "tqdm>=4.65.0",
    "datasets>=2.14.0",
    "sentence-transformers>=2.2.0",
    "ensemble-boxes>=1.0.7",
    "huggingface_hub>=0.16.0",
    "psutil>=5.9.5",
    "omegaconf",
    "pycocotools",
    "scipy>=1.11.4",
    "pyyaml",
    "fvcore>=0.1.5",
    "iopath>=0.1.7",
    "hydra-core>=1.3.2",
    "einops",
    "timm>=0.9.0",
    "spacy>=3.5.0",
    "nltk>=3.8.0",
    "blis>=0.7.11",
    "colorlog>=6.0.0",
    "pretty-errors>=1.2.25",
    "sentencepiece>=0.1.99",
    "num2words>=0.5.10",
]

[project.optional-dependencies]
segmentation = [
    "segment-anything-hq",
]
sam2 = [] # git dependency
detection = [
    "ultralytics>=8.0.0",
    # detectron2 requires separate install: pip install git+https://github.com/facebookresearch/detectron2.git
]
depth = [
    "depth-anything-v2",
]
vqa = [
    "accelerate>=0.20.0",
    "peft>=0.5.0",
    # vllm and ollama removed - users should install their preferred VLM library
]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "mypy>=1.0.0",
    "flake8>=6.0.0",
    "jupyter>=1.0.0",
    "ipykernel>=6.20.0",
]
visualization = [
    "adjustText>=0.7.3",
]
logging = [
    "wandb>=0.15.0",
]
all = [
    "segment-anything-hq",
    "ultralytics>=8.0.0",
    "depth-anything-v2",
    "accelerate>=0.20.0",
    "peft>=0.5.0",
    "adjustText>=0.7.3",
    "wandb>=0.15.0",
]
all-mps = [
    "segment-anything-hq",
    "ultralytics>=8.0.0",
    "depth-anything-v2",
    "accelerate>=0.20.0",
    "qwen-vl-utils>=0.0.10",
    "peft>=0.5.0",
    "adjustText>=0.7.3",
    "wandb>=0.15.0",
]

[project.urls]
Homepage = "https://github.com/disi-unibo-nlp/graph-of-marks"
Documentation = "https://github.com/disi-unibo-nlp/graph-of-marks#readme"
Repository = "https://github.com/disi-unibo-nlp/graph-of-marks"
"Bug Reports" = "https://github.com/disi-unibo-nlp/graph-of-marks/issues"

[project.scripts]
gom-preprocess = "gom.cli.preprocess:main"
gom-vqa = "gom.cli.vqa:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
include = ["gom*"]
exclude = ["tests*", "scripts*"]

[tool.setuptools.package-data]
gom = ["py.typed"]

[tool.black]
line-length = 100
target-version = ['py38', 'py39', 'py310', 'py311']
include = '\.pyi?$'

[tool.isort]
profile = "black"
line_length = 100
multi_line_output = 3

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
ignore_missing_imports = true

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-ra -q --strict-markers --cov=gom --cov-report=term-missing"
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]

[tool.uv.extra-build-dependencies]
detectron2 = ["torch>=2.4.0"]
